# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 5S Compliance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /submissions:
    get:
      operationId: listSubmissions
      tags: [submissions]
      summary: List submissions with optional filters
      parameters:
        - name: shift
          in: query
          schema:
            type: string
            enum: [A, B, C]
        - name: areaId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
    post:
      operationId: createSubmission
      tags: [submissions]
      summary: Submit a 5S video walk-through (preferred) or photo for scoring
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                areaId:
                  type: integer
                media:
                  type: string
                  format: binary
                  description: Video (preferred) or image file. Use this OR `photo`.
                photo:
                  type: string
                  format: binary
                  description: Legacy image-only field (still accepted).
                shift:
                  type: string
                  enum: [A, B, C]
                machineTag:
                  type: string
                  description: Optional sub-area / machine being captured.
              required:
                - areaId
      responses:
        "201":
          description: Submission created with score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/{id}:
    get:
      operationId: getSubmission
      tags: [submissions]
      summary: Get a single submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submission detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/{id}/reupload:
    put:
      operationId: reuploadSubmission
      tags: [submissions]
      summary: Re-upload media for an existing submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                media:
                  type: string
                  format: binary
                photo:
                  type: string
                  format: binary
                shift:
                  type: string
                  enum: [A, B, C]
                machineTag:
                  type: string
      responses:
        "200":
          description: Submission updated and re-scored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
...
  /operator/status:
    get:
      operationId: getOperatorStatus
      tags: [submissions]
      summary: Get operator submission status for current shift
      parameters:
        - name: shift
          in: query
          required: false
          schema:
            type: string
            enum: [A, B, C]
      responses:
        "200":
          description: Status per area for current shift
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AreaStatus"
...
components:
  schemas:
...
    Submission:
      type: object
      properties:
        id:
          type: integer
        areaId:
          type: integer
        areaName:
          type: string
        userId:
          type: integer
        userEmail:
          type: string
        shift:
          type: string
          enum: [A, B, C]
        scoreTotal:
          type: integer
        scoreJson:
          type: object
          properties:
            sort: { type: integer }
            set: { type: integer }
            shine: { type: integer }
            standardize: { type: integer }
            sustain: { type: integer }
          required: [sort, set, shine, standardize, sustain]
        suggestionsJson:
          type: array
          items:
            type: string
        imageUrl:
          type: string
        mediaType:
          type: string
          enum: [image, video]
        keyframesJson:
          type: array
          nullable: true
          items:
            type: string
        machineTag:
          type: string
          nullable: true
        failingPillarsJson:
          type: array
          nullable: true
          items:
            type: string
        aiTotalScore:
          type: integer
          nullable: true
        aiPillarsJson:
          type: object
          nullable: true
          properties:
            sort: { type: integer }
            set: { type: integer }
            shine: { type: integer }
            standardize: { type: integer }
            sustain: { type: integer }
        aiRecommendationsJson:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/AIRecommendation"
        aiIssuesJson:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/AIIssue"
        scoringMode:
          type: string
          nullable: true
        modelVersion:
          type: string
          nullable: true
        embeddingHash:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - areaId
        - areaName
        - userId
        - shift
        - scoreTotal
        - scoreJson
        - suggestionsJson
        - imageUrl
        - mediaType
        - createdAt
    AIRecommendation:
      type: object
      properties:
        action:
          type: string
        why:
          type: string
        location:
          type: string
        principle:
          type: string
          nullable: true
      required: [action, why, location]
    AIIssue:
      type: object
      properties:
        issue:
          type: string
        evidence:
          type: string
        location:
          type: string
        pillar:
          type: string
          nullable: true
        principle:
          type: string
          nullable: true
      required: [issue, evidence, location]
...
    AreaStatus:
      type: object
      properties:
        areaId:
          type: integer
        areaName:
          type: string
        submitted:
          type: boolean
        submission:
          $ref: "#/components/schemas/Submission"
      required: [areaId, areaName, submitted]
...